Emit code: split among new headers. Very limited logic changes.
