+++ src/c4/yml/emit.hpp
@@ -1,1131 +1,25 @@
 #ifndef _C4_YML_EMIT_HPP_
 #define _C4_YML_EMIT_HPP_
 
-/** @file emit.hpp Utilities to emit YAML and JSON. */
+/** @file emit.hpp Umbrella header with all utilities to emit YAML and JSON. */
 
-#ifndef _C4_YML_WRITER_HPP_
-#include "./writer.hpp"
+
+#ifndef _C4_YML_EMIT_BUF_HPP_
+#include "c4/yml/emit_buf.hpp"
+#endif
+#ifndef _C4_YML_EMIT_CONTAINER_HPP_
+#include "c4/yml/emit_container.hpp"
+#endif
+#ifndef _C4_YML_EMIT_FILE_HPP_
+#include "c4/yml/emit_file.hpp"
+#endif
+#ifndef _C4_YML_EMIT_OSTREAM_HPP_
+#include "c4/yml/emit_ostream.hpp"
 #endif
 
-#ifndef _C4_YML_TREE_HPP_
-#include "./tree.hpp"
-#endif
-
-#ifndef _C4_YML_NODE_HPP_
-#include "./node.hpp"
-#endif
-
-
-C4_SUPPRESS_WARNING_GCC_CLANG_WITH_PUSH("-Wold-style-cast")
-// NOLINTBEGIN(modernize-avoid-c-style-cast)
 
 namespace c4 {
 namespace yml {
-
-/** @cond dev */
-// fwd declarations
-template<class Writer> class Emitter;
-template<class OStream>
-using EmitterOStream = Emitter<WriterOStream<OStream>>;
-using EmitterFile = Emitter<WriterFile>;
-using EmitterBuf  = Emitter<WriterBuf>;
-/** Specifies the type of content to emit */
-typedef enum { // NOLINT
-    EMIT_YAML = 0, ///< emit YAML
-    EMIT_JSON = 1, ///< emit JSON
-} EmitType_e;
-/** @endcond */
-
-
-//-----------------------------------------------------------------------------
-//-----------------------------------------------------------------------------
-//-----------------------------------------------------------------------------
-
-/** A lightweight object containing options to be used when emitting.
- * @ingroup doc_emit
- * */
-struct EmitOptions
-{
-public:
-
-    /** @cond dev */
-    typedef enum : uint32_t { // NOLINT
-        INDENT_FLOW_ML = 1u << 0u,
-        FORCE_FLOW_SPC = 1u << 1u,
-        EMIT_NONROOT_KEY = 1u << 2u,
-        EMIT_NONROOT_DASH = 1u << 3u,
-        EMIT_NONROOT_MARKUP = EMIT_NONROOT_KEY|EMIT_NONROOT_DASH,
-        JSON_ERR_ON_STREAM = 1u << 4u,
-        JSON_ERR_ON_TAG = 1u << 5u,
-        JSON_ERR_ON_ANCHOR = 1u << 6u,
-        _JSON_ERR_MASK = JSON_ERR_ON_STREAM|JSON_ERR_ON_TAG|JSON_ERR_ON_ANCHOR,
-        DEFAULT_FLAGS = EMIT_NONROOT_KEY|INDENT_FLOW_ML,
-    } Flags_e;
-    /** @endcond */
-
-private:
-
-    EmitOptions& set_flags_(bool enabled, Flags_e f)
-    {
-        if(enabled)
-            m_flags |= f;
-        else
-            m_flags &= ~f;
-        return *this;
-    }
-
-public:
-
-    /** @name flow customization
-     *
-     * @{ */
-
-    /** Indent the contents of @ref FLOW_ML1 and @ref FLOW_MLN
-     * containers. Enabled by default. */
-    C4_ALWAYS_INLINE bool indent_flow_ml() const noexcept { return (m_flags & INDENT_FLOW_ML) != 0; }
-    EmitOptions& indent_flow_ml(bool enabled) noexcept { return set_flags_(enabled, INDENT_FLOW_ML); }
-
-    /** Force everywhere a space after comma in flow mode, overriding
-     * the @ref FLOW_SPC status of individual containers. This only
-     * applies to @ref FLOW_ML1 or @ref FLOW_MLN containers. Disabled
-     * by default. */
-    EmitOptions& force_flow_spc(bool enabled) noexcept { return set_flags_(enabled, FORCE_FLOW_SPC); }
-    C4_ALWAYS_INLINE bool force_flow_spc() const noexcept { return (m_flags & FORCE_FLOW_SPC) != 0; }
-
-    /** Set max columns for the emitted YAML in @ref FLOW_MLN
-     * mode. This will make the emitted YAML wrap around when the line
-     * reaches max cols, but only in containers with @ref FLOW_MLN
-     * mode. Defaults to 80 columns. */
-    EmitOptions& max_cols(id_type cols) noexcept { m_max_cols = cols; return *this; }
-    C4_ALWAYS_INLINE id_type max_cols() const noexcept { return m_max_cols; }
-    static constexpr const id_type max_cols_default = 80;
-
-    /** @} */
-
-public:
-
-    /** @name option flags - control emission of non-root (nested) nodes
-     *
-     * @{ */
-
-    /** When emit starts on a node which is not the root node, emit
-     * the node key as well. This will make the resuling YAML a map
-     * with the node as its single element. Enabled by default. */
-    EmitOptions& emit_nonroot_key(bool enabled) noexcept { return set_flags_(enabled, EMIT_NONROOT_KEY); }
-    C4_ALWAYS_INLINE bool emit_nonroot_key() const noexcept { return (m_flags & EMIT_NONROOT_KEY) != 0; }
-
-    /** When emit starts on a node which is not the root node, emit a
-     * leading dash. This will make the resulting YAML a seq with the
-     * node as its single element. Disabled by default. */
-    EmitOptions& emit_nonroot_dash(bool enabled) noexcept { return set_flags_(enabled, EMIT_NONROOT_DASH); }
-    C4_ALWAYS_INLINE bool emit_nonroot_dash() const noexcept { return (m_flags & EMIT_NONROOT_DASH) != 0; }
-
-    /** @} */
-
-public:
-
-    /** @name option flags - json behavior
-     *
-     * @{ */
-
-    /** Whether to trigger an error when findind a stream
-     * in json mode. Disabled by default. */
-    EmitOptions& json_err_on_stream(bool enabled) noexcept { return set_flags_(enabled, JSON_ERR_ON_STREAM); }
-    C4_ALWAYS_INLINE bool json_err_on_stream() const noexcept { return (m_flags & JSON_ERR_ON_STREAM) != 0; }
-
-    /** Whether to trigger an error (or ignore the tag) when finding a tag
-     * in json mode. Disabled by default. */
-    EmitOptions& json_err_on_tag(bool enabled) noexcept { return set_flags_(enabled, JSON_ERR_ON_TAG); }
-    C4_ALWAYS_INLINE bool json_err_on_tag() const noexcept { return (m_flags & JSON_ERR_ON_TAG) != 0; }
-
-    /** Whether to trigger an error (or ignore the anchor) when finding an
-     * anchor in json mode. Disabled by default. */
-    EmitOptions& json_err_on_anchor(bool enabled) noexcept { return set_flags_(enabled, JSON_ERR_ON_ANCHOR); }
-    C4_ALWAYS_INLINE bool json_err_on_anchor() const noexcept { return (m_flags & JSON_ERR_ON_ANCHOR) != 0; }
-
-    /** @cond dev */
-    RYML_DEPRECATED("use .json_err_on_{tag,anchor}()") C4_ALWAYS_INLINE Flags_e json_error_flags() const noexcept { return (Flags_e)(m_flags & _JSON_ERR_MASK); }
-    RYML_DEPRECATED("use .json_err_on_{tag,anchor}()") EmitOptions& json_error_flags(Flags_e d) noexcept { m_flags = (d & _JSON_ERR_MASK); return *this; }
-    /** @endcond */
-
-    /** @} */
-
-public:
-
-    /** @name maximum depth for the emitted tree
-     *
-     * This prevents stack overflows by making the emitter fail when
-     * the tree exceeds the maximum depth.
-     *
-     * @{ */
-    C4_ALWAYS_INLINE id_type max_depth() const noexcept { return m_max_depth; }
-    EmitOptions& max_depth(id_type d) noexcept { m_max_depth = d; return *this; }
-    static constexpr const id_type max_depth_default = 64;
-    /** @} */
-
-public:
-
-    bool operator== (const EmitOptions& that) const noexcept
-    {
-        return m_max_depth == that.m_max_depth &&
-            m_flags == that.m_flags;
-    }
-
-private:
-
-    /** @cond dev */
-    id_type m_max_depth{max_depth_default};
-    id_type m_max_cols{max_cols_default};
-    uint32_t m_flags{DEFAULT_FLAGS};
-    /** @endcond */
-};
-
-
-//-----------------------------------------------------------------------------
-//-----------------------------------------------------------------------------
-//-----------------------------------------------------------------------------
-
-/** A stateful emitter, for use with a writer such as @ref WriterBuf,
- * @ref WriterFile, or @ref WriterOStream
- * @ingroup doc_emit
- */
-template<class Writer>
-class Emitter : public Writer
-{
-public:
-
-    /** Construct the emitter and its internal Writer state.
-     *
-     * @param opts @ref EmitOptions
-     * @param args arguments to be forwarded to the constructor of the writer.
-     */
-    template<class ...WriterArgs>
-    Emitter(EmitOptions const& opts, WriterArgs &&...args) noexcept
-        : Writer(std::forward<WriterArgs>(args)...)
-        , m_tree()
-        , m_opts(opts)
-        , m_col()
-        , m_depth()
-        , m_ilevel()
-        , m_pws()
-        , m_flow_pws()
-    {}
-
-    /** Construct the emitter and its internal Writer state, using default emit options.
-     * @param args arguments to be forwarded to the constructor of the writer.
-     */
-    template<class ...WriterArgs>
-    Emitter(WriterArgs &&...args) noexcept
-        : Writer(std::forward<WriterArgs>(args)...)
-        , m_tree()
-        , m_opts()
-        , m_col()
-        , m_depth()
-        , m_ilevel()
-        , m_pws()
-        , m_flow_pws()
-    {}
-
-public:
-
-    /** emit!
-     *
-     * When writing to a buffer, returns a substr of the emitted YAML.
-     * If the given buffer has insufficient space, the returned substr
-     * will be null and its size will be the needed space. Whatever
-     * the size of the buffer, it is guaranteed that no writes are
-     * done past its end.
-     *
-     * When writing to a file, the returned substr will be null, but its
-     * length will be set to the number of bytes written.
-     *
-     * @param type specify what to emit (YAML or JSON)
-     * @param t the tree to emit
-     * @param id the id of the node to emit
-     */
-    void emit_as(EmitType_e type, Tree const& t, id_type id);
-    /** emit starting at the root node */
-    void emit_as(EmitType_e type, Tree const& t)
-    {
-        return this->emit_as(type, t, t.root_id());
-    }
-    /** emit starting at the given node */
-    void emit_as(EmitType_e type, ConstNodeRef const& n)
-    {
-        if(!n.readable())
-            return;
-        return this->emit_as(type, *n.tree(), n.id());
-    }
-
-public:
-
-    /** get the emit options for this object */
-    EmitOptions const& options() const noexcept { return m_opts; }
-    /** set the emit options for this object */
-    void options(EmitOptions opts) noexcept { m_opts = opts; }
-
-    /** set the max depth for emitted trees (to prevent a stack overflow) */
-    void max_depth(id_type max_depth) noexcept { m_opts.max_depth(max_depth); }
-    /** get the max depth for emitted trees (to prevent a stack overflow) */
-    id_type max_depth() const noexcept { return m_opts.max_depth(); }
-
-private: // pending whitespace
-
-    /// pending whitespace
-    typedef enum : uint32_t { _PWS_NONE = 0u, _PWS_SPACE = 1u, _PWS_NEWL = 2u } Pws_e; // NOLINT
-
-    /// set pending whitespace, ignoring pending
-    C4_ALWAYS_INLINE void _pend_none() noexcept
-    {
-        m_pws = _PWS_NONE;
-    }
-    /// set pending whitespace, ignoring pending
-    C4_ALWAYS_INLINE void _pend_newl() noexcept
-    {
-        m_pws = _PWS_NEWL;
-    }
-    /// set pending whitespace, ignoring pending
-    C4_ALWAYS_INLINE void _pend_space() noexcept
-    {
-        m_pws = _PWS_SPACE;
-    }
-    /// write pending whitespace, and then set the next pending whitespace
-    C4_ALWAYS_INLINE void _write_pws_and_pend(Pws_e next=_PWS_NONE) noexcept
-    {
-        if(m_pws == _PWS_SPACE)
-        {
-            _write(' ');
-        }
-        else if(m_pws == _PWS_NEWL)
-        {
-            _newl();
-            _indent(m_ilevel);
-        }
-        m_pws = next;
-    }
-
-    /// specs for obtaining pending whitespace in flow mode
-    struct flow_pws
-    {
-        size_t max_cols = 0; // leave this member first to avoid padding
-        Pws_e pend_after_comma = _PWS_NONE;
-        bool active = false;
-        C4_ALWAYS_INLINE Pws_e next_pws(size_t col) const noexcept
-        {
-            return (active && col >= max_cols) ? _PWS_NEWL : pend_after_comma;
-        }
-        void start(NodeType ty, size_t max_cols_) noexcept;
-        void stop() noexcept { active = false; }
-    };
-
-    C4_NODISCARD bool _maybe_start_flow_pws_ml(id_type node) noexcept
-    {
-        _RYML_ASSERT_VISIT_(m_tree->callbacks(), m_tree->type(node) & (FLOW_ML1|FLOW_MLN), m_tree, node);
-        if(m_flow_pws.active)
-            return false;
-        NodeType ty = m_tree->type(node);
-        if(m_opts.force_flow_spc())
-            ty |= FLOW_SPC;
-        m_flow_pws.start(ty, m_opts.max_cols());
-        return true;
-    }
-    C4_NODISCARD flow_pws _setup_flow_pws_sl(id_type node) noexcept
-    {
-        flow_pws ret = {};
-        if(m_flow_pws.active)
-        {
-            ret = m_flow_pws;
-        }
-        else
-        {
-            NodeType ty = m_tree->type(node);
-            if(m_opts.force_flow_spc())
-                ty |= FLOW_SPC;
-            ret.start(ty, 0);
-        }
-        return ret;
-    }
-
-private:
-
-    /** @cond dev */
-
-    void _emit_yaml(id_type id);
-
-    void _visit_stream(id_type id);
-    void _visit_doc(id_type id);
-    void _visit_doc_val(id_type id);
-    void _visit_blck_container(id_type id);
-    void _visit_flow_container(id_type id);
-
-    void _visit_flow_sl(id_type id);
-    void _visit_flow_sl_seq(id_type id);
-    void _visit_flow_sl_map(id_type id);
-
-    void _visit_flow_ml(id_type id);
-    void _visit_flow_ml_seq(id_type id);
-    void _visit_flow_ml_map(id_type id);
-
-    void _visit_blck(id_type id);
-    void _visit_blck_seq(id_type id);
-    void _visit_blck_map(id_type id);
-
-    void _top_open_entry(id_type id);
-    void _top_close_entry(id_type id);
-    void _blck_seq_open_entry(id_type id);
-    void _blck_map_open_entry(id_type id);
-    void _blck_close_entry(id_type id);
-    void _blck_write_scalar(csubstr str, type_bits type);
-
-    void _flow_seq_open_entry(id_type id);
-    void _flow_map_open_entry(id_type id);
-    void _flow_close_entry_sl(id_type id, id_type last_sibling, Pws_e pend_after);
-    void _flow_close_entry_ml(id_type id, id_type last_sibling, Pws_e pend_after);
-    void _flow_write_scalar(csubstr str, type_bits type);
-
-private:
-
-    void _json_emit(id_type id);
-    void _write_scalar_literal(csubstr s, id_type level);
-    void _write_scalar_folded(csubstr s, id_type level);
-    void _write_scalar_squo(csubstr s, id_type level);
-    void _write_scalar_dquo(csubstr s, id_type level);
-    void _write_scalar_plain(csubstr s, id_type level);
-
-    size_t _write_escaped_newlines(csubstr s, size_t i);
-    size_t _write_indented_block(csubstr s, size_t i, id_type level);
-
-private:
-
-    void _json_visit_ml(id_type id, NodeType ty, id_type depth);
-    void _json_visit_sl(id_type id, NodeType ty, id_type depth);
-    bool _json_maybe_write_naninf(csubstr s);
-    void _json_writek(id_type id, NodeType ty);
-    void _json_writev(id_type id, NodeType ty);
-    void _json_write_scalar_dquo(csubstr s);
-    void _json_write_number(csubstr s);
-
-private:
-
-    void _write_tag(csubstr tag)
-    {
-        if(!tag.begins_with('!'))
-            _write('!');
-        _write(tag);
-    }
-    void _write_ref(csubstr ref)
-    {
-        if(ref != "<<")
-        {
-            if(!ref.begins_with('*'))
-                _write('*');
-            _write(ref);
-        }
-    }
-
-private:
-
-    C4_ALWAYS_INLINE void _indent(id_type level)
-    {
-        size_t num = (size_t)(2u * level);
-        this->Writer::_do_write(' ', num);
-        m_col += num;
-    }
-
-    template<size_t N>
-    C4_ALWAYS_INLINE void _write(const char (&a)[N])
-    {
-        this->Writer::_do_write(std::forward<const char (&)[N]>(a));
-        m_col += N-1;
-    }
-    C4_ALWAYS_INLINE void _write(csubstr s)
-    {
-        this->Writer::_do_write(s);
-        m_col += s.len;
-    }
-    C4_ALWAYS_INLINE void _write(char c)
-    {
-        this->Writer::_do_write(c);
-        ++m_col;
-    }
-    C4_ALWAYS_INLINE void _write(char c, size_t num)
-    {
-        this->Writer::_do_write(c, num);
-        m_col += num;
-    }
-
-    /// write a newline and reset the column
-    C4_ALWAYS_INLINE void _newl()
-    {
-        this->Writer::_do_write('\n');
-        m_col = 0;
-    }
-
-private:
-
-    Tree const* C4_RESTRICT m_tree;
-    EmitOptions m_opts;
-    size_t      m_col;
-    id_type     m_depth;
-    id_type     m_ilevel;
-    Pws_e       m_pws;
-    flow_pws    m_flow_pws;
-
-private:
-
-    C4_SUPPRESS_WARNING_GCC_PUSH
-    #if defined(__GNUC__) && (__GNUC__ < 5) && (!defined(__clang__))
-    // g++-4.x has problems with the operand types here...
-    C4_SUPPRESS_WARNING_GCC_WITH_PUSH("-Wparentheses")
-    #endif
-    enum : type_bits { // NOLINT
-        _styles_block_key = KEY_LITERAL|KEY_FOLDED,
-        _styles_block_val = VAL_LITERAL|VAL_FOLDED,
-        _styles_block     = ((type_bits)_styles_block_key) | ((type_bits)_styles_block_val),
-        _styles_flow_key  = KEY_STYLE & (~((type_bits)_styles_block_key)),
-        _styles_flow_val  = VAL_STYLE & (~((type_bits)_styles_block_val)),
-        _styles_flow      = ((type_bits)_styles_flow_key) | ((type_bits)_styles_flow_val),
-        _styles_squo      = KEY_SQUO|VAL_SQUO,
-        _styles_dquo      = KEY_DQUO|VAL_DQUO,
-        _styles_plain     = KEY_PLAIN|VAL_PLAIN,
-        _styles_literal   = KEY_LITERAL|VAL_LITERAL,
-        _styles_folded    = KEY_FOLDED|VAL_FOLDED,
-    };
-    C4_SUPPRESS_WARNING_GCC_POP
-
-    /** @endcond */
-};
-
-
-//-----------------------------------------------------------------------------
-//-----------------------------------------------------------------------------
-//-----------------------------------------------------------------------------
-
-// emit from tree and node id -----------------------
-
-/** @addtogroup doc_emit_to_file_from_node_id
- *
- * @{
- */
-
-/** (1) emit YAML to the given file, starting at the given node. A null
- * file defaults to stdout. */
-inline void emit_yaml(Tree const& t, id_type id, EmitOptions const& opts, FILE *f)
-{
-    EmitterFile em(opts, f);
-    em.emit_as(EMIT_YAML, t, id);
-}
-/** (2) like (1), but use default emit options */
-inline void emit_yaml(Tree const& t, id_type id, FILE *f)
-{
-    EmitterFile em(f);
-    em.emit_as(EMIT_YAML, t, id);
-}
-
-/** (1) emit JSON to the given file, starting at the given node. A null
- * file defaults to stdout. */
-inline void emit_json(Tree const& t, id_type id, EmitOptions const& opts, FILE *f)
-{
-    EmitterFile em(opts, f);
-    em.emit_as(EMIT_JSON, t, id);
-}
-/** (2) like (1), but use default emit options */
-inline void emit_json(Tree const& t, id_type id, FILE *f)
-{
-    EmitterFile em(f);
-    em.emit_as(EMIT_JSON, t, id);
-}
-
-/** @} */
-
-
-// emit from root -------------------------
-
-/** @addtogroup doc_emit_to_file_from_root
- *
- * @{
- */
-
-/** (1) emit YAML to the given file, starting at the root node. A null file defaults to stdout. */
-inline void emit_yaml(Tree const& t, EmitOptions const& opts, FILE *f=nullptr)
-{
-    EmitterFile em(opts, f);
-    em.emit_as(EMIT_YAML, t);
-}
-/** (2) like (1), but use default emit options */
-inline void emit_yaml(Tree const& t, FILE *f=nullptr)
-{
-    EmitterFile em(f);
-    em.emit_as(EMIT_YAML, t);
-}
-
-/** (1) emit JSON to the given file. A null file defaults to stdout. */
-inline void emit_json(Tree const& t, EmitOptions const& opts, FILE *f=nullptr)
-{
-    EmitterFile em(opts, f);
-    em.emit_as(EMIT_JSON, t);
-}
-/** (2) like (1), but use default emit options */
-inline void emit_json(Tree const& t, FILE *f=nullptr)
-{
-    EmitterFile em(f);
-    em.emit_as(EMIT_JSON, t);
-}
-
-/** @} */
-
-
-// emit from ConstNodeRef ------------------------
-
-/** @addtogroup doc_emit_to_file_from_noderef
- *
- * @{
- */
-
-/** (1) emit YAML to the given file. A null file defaults to stdout.
- * Return the number of bytes written. */
-inline void emit_yaml(ConstNodeRef const& r, EmitOptions const& opts, FILE *f=nullptr)
-{
-    if(!r.readable())
-        return;
-    EmitterFile em(opts, f);
-    em.emit_as(EMIT_YAML, r);
-}
-/** (2) like (1), but use default emit options */
-inline void emit_yaml(ConstNodeRef const& r, FILE *f=nullptr)
-{
-    if(!r.readable())
-        return;
-    EmitterFile em(f);
-    em.emit_as(EMIT_YAML, r);
-}
-
-/** (1) emit JSON to the given file. A null file defaults to stdout.
- * Return the number of bytes written. */
-inline void emit_json(ConstNodeRef const& r, EmitOptions const& opts, FILE *f=nullptr)
-{
-    if(!r.readable())
-        return;
-    EmitterFile em(opts, f);
-    em.emit_as(EMIT_JSON, r);
-}
-/** (2) like (1), but use default emit options */
-inline void emit_json(ConstNodeRef const& r, FILE *f=nullptr)
-{
-    if(!r.readable())
-        return;
-    EmitterFile em(f);
-    em.emit_as(EMIT_JSON, r);
-}
-
-/** @} */
-
-
-//-----------------------------------------------------------------------------
-
-/** @addtogroup doc_emit_to_ostream
- *
- * @{
- */
-
-
-/** emit YAML to an STL-like ostream */
-template<class OStream>
-inline OStream& operator<< (OStream& s, Tree const& t)
-{
-    EmitterOStream<OStream> em(s);
-    em.emit_as(EMIT_YAML, t);
-    return s;
-}
-
-/** emit YAML to an STL-like ostream
- * @overload */
-template<class OStream>
-inline OStream& operator<< (OStream& s, ConstNodeRef const& n)
-{
-    if(!n.readable())
-        return s;
-    EmitterOStream<OStream> em(s);
-    em.emit_as(EMIT_YAML, n);
-    return s;
-}
-
-/** tag type to mark a tree or node to be emitted as yaml when using
- * @ref operator<<, with options. For example:
- *
- * ```cpp
- * Tree t = parse_in_arena("{foo: bar}");
- * std::cout << t; // emits YAML
- * std::cout << as_yaml(t); // emits YAML, same as above
- * std::cout << as_yaml(t, EmitOptions{}.flow_spc(true)); // emits YAML forcing spaces after flow commas
- * ```
- *
- * @see @ref operator<< */
-struct as_json
-{
-    Tree const* tree;
-    id_type node;
-    EmitOptions options;
-    as_json(Tree const& t, EmitOptions const& opts={}) : tree(&t), node(t.empty() ? NONE : t.root_id()), options(opts)  {}
-    as_json(Tree const& t, id_type id, EmitOptions const& opts={}) : tree(&t), node(id), options(opts)  {}
-    as_json(ConstNodeRef const& n, EmitOptions const& opts={}) : tree(n.tree()), node(n.id()), options(opts) {}
-};
-
-/** tag type to mark a tree or node to be emitted as yaml when using
- * @ref operator<< . For example:
- *
- * ```cpp
- * Tree t = parse_in_arena("{foo: bar}");
- * std::cout << t; // emits YAML
- * std::cout << as_json(t); // emits JSON
- * std::cout << as_json(t, EmitOptions{}.flow_spc(true)); // emits JSON forcing spaces after flow commas
- * ```
- *
- * @see @ref operator<< */
-struct as_yaml
-{
-    Tree const* tree;
-    id_type node;
-    EmitOptions options;
-    as_yaml(Tree const& t, EmitOptions const& opts={}) : tree(&t), node(t.empty() ? NONE : t.root_id()), options(opts)  {}
-    as_yaml(Tree const& t, id_type id, EmitOptions const& opts={}) : tree(&t), node(id), options(opts)  {}
-    as_yaml(ConstNodeRef const& n, EmitOptions const& opts={}) : tree(n.tree()), node(n.id()), options(opts) {}
-};
-
-/** emit json to an STL-like stream */
-template<class OStream>
-inline OStream& operator<< (OStream& s, as_json const& j)
-{
-    if(!j.tree || j.tree->empty())
-        return s;
-    EmitterOStream<OStream> em(j.options, s);
-    em.emit_as(EMIT_JSON, *j.tree, j.node != NONE ? j.node : j.tree->root_id());
-    return s;
-}
-
-/** emit yaml to an STL-like stream */
-template<class OStream>
-inline OStream& operator<< (OStream& s, as_yaml const& y)
-{
-    if(!y.tree || y.tree->empty())
-        return s;
-    EmitterOStream<OStream> em(y.options, s);
-    em.emit_as(EMIT_YAML, *y.tree, y.node != NONE ? y.node : y.tree->root_id());
-    return s;
-}
-
-/** @} */
-
-
-//-----------------------------------------------------------------------------
-
-/** @addtogroup doc_emit_to_buffer_from_node_id
- *
- * @{
- */
-
-// emit from tree and node id -----------------------
-
-/** (1) emit YAML to the given buffer. Return a substr trimmed to the emitted YAML.
- * @param t the tree to emit.
- * @param id the node where to start emitting.
- * @param opts emit options.
- * @param buf the output buffer.
- * @param error_on_excess Raise an error if the space in the buffer is insufficient.
- * @return a substr trimmed to the result in the output buffer. If the buffer is
- * insufficient (when error_on_excess is false), the string pointer of the
- * result will be set to null, and the length will report the required buffer size. */
-inline substr emit_yaml(Tree const& t, id_type id, EmitOptions const& opts, substr buf, bool error_on_excess=true)
-{
-    EmitterBuf em(opts, buf);
-    em.emit_as(EMIT_YAML, t, id);
-    return em._get(error_on_excess);
-}
-/** (2) like (1), but use default emit options */
-inline substr emit_yaml(Tree const& t, id_type id, substr buf, bool error_on_excess=true)
-{
-    EmitterBuf em(buf);
-    em.emit_as(EMIT_YAML, t, id);
-    return em._get(error_on_excess);
-}
-
-/** (1) emit JSON to the given buffer. Return a substr trimmed to the emitted JSON.
- * @param t the tree to emit.
- * @param id the node where to start emitting.
- * @param opts emit options.
- * @param buf the output buffer.
- * @param error_on_excess Raise an error if the space in the buffer is insufficient.
- * @return a substr trimmed to the result in the output buffer. If the buffer is
- * insufficient (when error_on_excess is false), the string pointer of the
- * result will be set to null, and the length will report the required buffer size. */
-inline substr emit_json(Tree const& t, id_type id, EmitOptions const& opts, substr buf, bool error_on_excess=true)
-{
-    EmitterBuf em(opts, buf);
-    em.emit_as(EMIT_JSON, t, id);
-    return em._get(error_on_excess);
-}
-/** (2) like (1), but use default emit options */
-inline substr emit_json(Tree const& t, id_type id, substr buf, bool error_on_excess=true)
-{
-    EmitterBuf em(buf);
-    em.emit_as(EMIT_JSON, t, id);
-    return em._get(error_on_excess);
-}
-
-/** @} */
-
-
-// emit from root -------------------------
-
-/** @addtogroup doc_emit_to_buffer_from_root
- *
- * @{
- */
-
-/** (1) emit YAML to the given buffer. Return a substr trimmed to the emitted YAML.
- * @param t the tree; will be emitted from the root node.
- * @param opts emit options.
- * @param buf the output buffer.
- * @param error_on_excess Raise an error if the space in the buffer is insufficient.
- * @return a substr trimmed to the result in the output buffer. If the buffer is
- * insufficient (when error_on_excess is false), the string pointer of the
- * result will be set to null, and the length will report the required buffer size. */
-inline substr emit_yaml(Tree const& t, EmitOptions const& opts, substr buf, bool error_on_excess=true)
-{
-    EmitterBuf em(opts, buf);
-    em.emit_as(EMIT_YAML, t);
-    return em._get(error_on_excess);
-}
-/** (2) like (1), but use default emit options */
-inline substr emit_yaml(Tree const& t, substr buf, bool error_on_excess=true)
-{
-    EmitterBuf em(buf);
-    em.emit_as(EMIT_YAML, t);
-    return em._get(error_on_excess);
-}
-
-/** (1) emit JSON to the given buffer. Return a substr trimmed to the emitted JSON.
- * @param t the tree; will be emitted from the root node.
- * @param opts emit options.
- * @param buf the output buffer.
- * @param error_on_excess Raise an error if the space in the buffer is insufficient.
- * @return a substr trimmed to the result in the output buffer. If the buffer is
- * insufficient (when error_on_excess is false), the string pointer of the
- * result will be set to null, and the length will report the required buffer size. */
-inline substr emit_json(Tree const& t, EmitOptions const& opts, substr buf, bool error_on_excess=true)
-{
-    EmitterBuf em(opts, buf);
-    em.emit_as(EMIT_JSON, t);
-    return em._get(error_on_excess);
-}
-/** (2) like (1), but use default emit options */
-inline substr emit_json(Tree const& t, substr buf, bool error_on_excess=true)
-{
-    EmitterBuf em(buf);
-    em.emit_as(EMIT_JSON, t);
-    return em._get(error_on_excess);
-}
-
-/** @} */
-
-
-// emit from ConstNodeRef ------------------------
-
-/** @addtogroup doc_emit_to_buffer_from_noderef
- *
- * @{
- */
-
-/** (1) emit YAML to the given buffer. Return a substr trimmed to the emitted YAML.
- * @param r the starting node.
- * @param buf the output buffer.
- * @param opts emit options.
- * @param error_on_excess Raise an error if the space in the buffer is insufficient.
- * @return a substr trimmed to the result in the output buffer. If the buffer is
- * insufficient (when error_on_excess is false), the string pointer of the
- * result will be set to null, and the length will report the required buffer size. */
-inline substr emit_yaml(ConstNodeRef const& r, EmitOptions const& opts, substr buf, bool error_on_excess=true)
-{
-    if(!r.readable())
-        return {};
-    EmitterBuf em(opts, buf);
-    em.emit_as(EMIT_YAML, r);
-    return em._get(error_on_excess);
-}
-/** (2) like (1), but use default emit options */
-inline substr emit_yaml(ConstNodeRef const& r, substr buf, bool error_on_excess=true)
-{
-    if(!r.readable())
-        return {};
-    EmitterBuf em(buf);
-    em.emit_as(EMIT_YAML, r);
-    return em._get(error_on_excess);
-}
-
-/** (1) emit JSON to the given buffer. Return a substr trimmed to the emitted JSON.
- * @param r the starting node.
- * @param buf the output buffer.
- * @param opts emit options.
- * @param error_on_excess Raise an error if the space in the buffer is insufficient.
- * @return a substr trimmed to the result in the output buffer. If the buffer is
- * insufficient (when error_on_excess is false), the string pointer of the
- * result will be set to null, and the length will report the required buffer size. */
-inline substr emit_json(ConstNodeRef const& r, EmitOptions const& opts, substr buf, bool error_on_excess=true)
-{
-    if(!r.readable())
-        return {};
-    EmitterBuf em(opts, buf);
-    em.emit_as(EMIT_JSON, r);
-    return em._get(error_on_excess);
-}
-/** (2) like (1), but use default emit options */
-inline substr emit_json(ConstNodeRef const& r, substr buf, bool error_on_excess=true)
-{
-    if(!r.readable())
-        return {};
-    EmitterBuf em(buf);
-    em.emit_as(EMIT_JSON, r);
-    return em._get(error_on_excess);
-}
-
-/** @} */
-
-
-//-----------------------------------------------------------------------------
-
-// emit from tree and node id ---------------------------
-
-/** @addtogroup doc_emit_to_container_from_node_id
- *
- * @{
- */
-
-/** (1) emit+resize: emit YAML to the given `std::string`/`std::vector`-like
- * container, resizing it as needed to fit the emitted YAML. If @p append is
- * set to true, the emitted YAML is appended at the end of the container.
- *
- * @return a substr trimmed to the emitted YAML (excluding the initial contents, when appending) */
-template<class CharOwningContainer>
-substr emitrs_yaml(Tree const& t, id_type id, EmitOptions const& opts, CharOwningContainer * cont, bool append=false)
-{
-    size_t startpos = append ? cont->size() : 0u;
-    cont->resize(cont->capacity()); // otherwise the first emit would be certain to fail
-    substr buf = to_substr(*cont).sub(startpos);
-    substr ret = emit_yaml(t, id, opts, buf, /*error_on_excess*/false);
-    if(ret.str == nullptr && ret.len > 0)
-    {
-        cont->resize(startpos + ret.len);
-        buf = to_substr(*cont).sub(startpos);
-        ret = emit_yaml(t, id, opts, buf, /*error_on_excess*/true);
-    }
-    else
-    {
-        cont->resize(startpos + ret.len);
-    }
-    return ret;
-}
-/** (2) like (1), but use default emit options */
-template<class CharOwningContainer>
-substr emitrs_yaml(Tree const& t, id_type id, CharOwningContainer * cont, bool append=false)
-{
-    return emitrs_yaml(t, id, EmitOptions{}, cont, append);
-}
-
-/** (1) emit+resize: emit JSON to the given `std::string`/`std::vector`-like
- * container, resizing it as needed to fit the emitted JSON. If @p append is
- * set to true, the emitted YAML is appended at the end of the container.
- *
- * @return a substr trimmed to the emitted JSON (excluding the initial contents, when appending) */
-template<class CharOwningContainer>
-substr emitrs_json(Tree const& t, id_type id, EmitOptions const& opts, CharOwningContainer * cont, bool append=false)
-{
-    const size_t startpos = append ? cont->size() : 0u;
-    cont->resize(cont->capacity()); // otherwise the first emit would be certain to fail
-    substr buf = to_substr(*cont).sub(startpos);
-    EmitterBuf em(opts, buf);
-    substr ret = emit_json(t, id, opts, buf, /*error_on_excess*/false);
-    if(ret.str == nullptr && ret.len > 0)
-    {
-        cont->resize(startpos + ret.len);
-        buf = to_substr(*cont).sub(startpos);
-        ret = emit_json(t, id, opts, buf, /*error_on_excess*/true);
-    }
-    else
-    {
-        cont->resize(startpos + ret.len);
-    }
-    return ret;
-}
-/** (2) like (1), but use default emit options */
-template<class CharOwningContainer>
-substr emitrs_json(Tree const& t, id_type id, CharOwningContainer * cont, bool append=false)
-{
-    return emitrs_json(t, id, EmitOptions{}, cont, append);
-}
-
-
-/** (3) emit+resize: YAML to a newly-created `std::string`/`std::vector`-like container. */
-template<class CharOwningContainer>
-CharOwningContainer emitrs_yaml(Tree const& t, id_type id, EmitOptions const& opts={})
-{
-    CharOwningContainer c;
-    emitrs_yaml(t, id, opts, &c);
-    return c;
-}
-/** (3) emit+resize: JSON to a newly-created `std::string`/`std::vector`-like container. */
-template<class CharOwningContainer>
-CharOwningContainer emitrs_json(Tree const& t, id_type id, EmitOptions const& opts={})
-{
-    CharOwningContainer c;
-    emitrs_json(t, id, opts, &c);
-    return c;
-}
-
-/** @} */
-
-
-// emit from root -------------------------
-
-/** @addtogroup doc_emit_to_container_from_root
- *
- * @{
- */
-
-/** (1) emit+resize: YAML to the given `std::string`/`std::vector`-like
- * container, resizing it as needed to fit the emitted YAML.
- * @return a substr trimmed to the new emitted contents. */
-template<class CharOwningContainer>
-substr emitrs_yaml(Tree const& t, EmitOptions const& opts, CharOwningContainer * cont, bool append=false)
-{
-    if(t.empty())
-        return {};
-    return emitrs_yaml(t, t.root_id(), opts, cont, append);
-}
-/** (2) like (1), but use default emit options */
-template<class CharOwningContainer>
-substr emitrs_yaml(Tree const& t, CharOwningContainer * cont, bool append=false)
-{
-    if(t.empty())
-        return {};
-    return emitrs_yaml
-        (t, t.root_id(), EmitOptions{}, cont, append);
-}
-
-/** (1) emit+resize: JSON to the given `std::string`/`std::vector`-like
- * container, resizing it as needed to fit the emitted JSON.
- * @return a substr trimmed to the new emitted contents. */
-template<class CharOwningContainer>
-substr emitrs_json(Tree const& t, EmitOptions const& opts, CharOwningContainer * cont, bool append=false)
-{
-    if(t.empty())
-        return {};
-    return emitrs_json(t, t.root_id(), opts, cont, append);
-}
-/** (2) like (1), but use default emit options */
-template<class CharOwningContainer>
-substr emitrs_json(Tree const& t, CharOwningContainer * cont, bool append=false)
-{
-    if(t.empty())
-        return {};
-    return emitrs_json(t, t.root_id(), EmitOptions{}, cont, append);
-}
-
-
-/** (3) emit+resize: YAML to a newly-created `std::string`/`std::vector`-like container. */
-template<class CharOwningContainer>
-CharOwningContainer emitrs_yaml(Tree const& t, EmitOptions const& opts={})
-{
-    CharOwningContainer c;
-    if(t.empty())
-        return c;
-    emitrs_yaml(t, t.root_id(), opts, &c);
-    return c;
-}
-/** (3) emit+resize: JSON to a newly-created `std::string`/`std::vector`-like container. */
-template<class CharOwningContainer>
-CharOwningContainer emitrs_json(Tree const& t, EmitOptions const& opts={})
-{
-    CharOwningContainer c;
-    if(t.empty())
-        return c;
-    emitrs_json(t, t.root_id(), opts, &c);
-    return c;
-}
-
-/** @} */
-
-
-// emit from ConstNodeRef ------------------------
-
-/** @addtogroup doc_emit_to_container_from_noderef
- *
- * @{
- */
-
-/** (1) emit+resize: YAML to the given `std::string`/`std::vector`-like container,
- * resizing it as needed to fit the emitted YAML.
- * @return a substr trimmed to the new emitted contents */
-template<class CharOwningContainer>
-substr emitrs_yaml(ConstNodeRef const& n, EmitOptions const& opts, CharOwningContainer * cont, bool append=false)
-{
-    if(!n.readable())
-        return {};
-    return emitrs_yaml(*n.tree(), n.id(), opts, cont, append);
-}
-/** (2) like (1), but use default emit options */
-template<class CharOwningContainer>
-substr emitrs_yaml(ConstNodeRef const& n, CharOwningContainer * cont, bool append=false)
-{
-    if(!n.readable())
-        return {};
-    return emitrs_yaml(*n.tree(), n.id(), EmitOptions{}, cont, append);
-}
-
-/** (1) emit+resize: JSON to the given `std::string`/`std::vector`-like container,
- * resizing it as needed to fit the emitted JSON.
- * @return a substr trimmed to the new emitted contents */
-template<class CharOwningContainer>
-substr emitrs_json(ConstNodeRef const& n, EmitOptions const& opts, CharOwningContainer * cont, bool append=false)
-{
-    if(!n.readable())
-        return {};
-    return emitrs_json(*n.tree(), n.id(), opts, cont, append);
-}
-/** (2) like (1), but use default emit options */
-template<class CharOwningContainer>
-substr emitrs_json(ConstNodeRef const& n, CharOwningContainer * cont, bool append=false)
-{
-    if(!n.readable())
-        return {};
-    return emitrs_json(*n.tree(), n.id(), EmitOptions{}, cont, append);
-}
-
-
-/** (3) emit+resize: YAML to a newly-created `std::string`/`std::vector`-like container. */
-template<class CharOwningContainer>
-CharOwningContainer emitrs_yaml(ConstNodeRef const& n, EmitOptions const& opts={})
-{
-    if(!n.readable())
-        return {};
-    CharOwningContainer c;
-    emitrs_yaml(*n.tree(), n.id(), opts, &c);
-    return c;
-}
-/** (3) emit+resize: JSON to a newly-created `std::string`/`std::vector`-like container. */
-template<class CharOwningContainer>
-CharOwningContainer emitrs_json(ConstNodeRef const& n, EmitOptions const& opts={})
-{
-    if(!n.readable())
-        return {};
-    CharOwningContainer c;
-    emitrs_json(*n.tree(), n.id(), opts, &c);
-    return c;
-}
-
-/** @} */
-
-
-//-----------------------------------------------------------------------------
 
 /** @cond dev */
 
@@ -1206,18 +100,13 @@ RYML_DEPRECATE_EMITRS CharOwningContainer emitrs(ConstNodeRef const& n)
     return emitrs_yaml<CharOwningContainer>(n);
 }
 
+#undef RYML_DEPRECATE_EMIT
+#undef RYML_DEPRECATE_EMITRS
+
 /** @endcond */
 
 
 } // namespace yml
 } // namespace c4
-
-// NOLINTEND(modernize-avoid-c-style-cast)
-C4_SUPPRESS_WARNING_GCC_CLANG_POP
-
-#undef RYML_DEPRECATE_EMIT
-#undef RYML_DEPRECATE_EMITRS
-
-#include "c4/yml/emit.def.hpp" // NOLINT
 
 #endif /* _C4_YML_EMIT_HPP_ */
