+++ tools/amalgamate.py
@@ -135,11 +135,20 @@ INSTRUCTIONS:
         "src/c4/yml/tag.hpp",
         am.onlyif(has_evt(Event.tree), "src/c4/yml/tree.hpp"),
         am.onlyif(has_evt(Event.tree), "src/c4/yml/node.hpp"),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/writer_buf.hpp"),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/writer_file.hpp"),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/writer_ostream.hpp"),
         am.onlyif(has_evt(Event.tree), "src/c4/yml/writer.hpp"),
         "src/c4/yml/detail/dbgprint.hpp",
         am.onlyif(has_evt(Event.tree), am.injcode("#define C4_YML_EMIT_DEF_HPP_")),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/emit_options.hpp"),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/emitter.hpp"),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/emitter.def.hpp"),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/emit_buf.hpp"),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/emit_container.hpp"),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/emit_file.hpp"),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/emit_ostream.hpp"),
         am.onlyif(has_evt(Event.tree), "src/c4/yml/emit.hpp"),
-        am.onlyif(has_evt(Event.tree), "src/c4/yml/emit.def.hpp"),
         "src/c4/yml/filter_processor.hpp",
         "src/c4/yml/parser_state.hpp",
         "src/c4/yml/event_handler_stack.hpp",
@@ -166,6 +175,8 @@ INSTRUCTIONS:
         am.onlyif(has_evt(Event.ints_to_testsuite), "src_extra/c4/yml/extra/ints_to_testsuite.cpp"),
         am.onlyif(has_evt(Event.tree), "src/c4/yml/reference_resolver.cpp"),
         am.onlyif(has_evt(Event.tree), "src/c4/yml/parse.cpp"),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/emit_buf.cpp"),
+        am.onlyif(has_evt(Event.tree), "src/c4/yml/emit_file.cpp"),
         "src/c4/yml/preprocess.cpp",
         am.onlyif(has_evt(Event.tree), "src/c4/yml/detail/checks.hpp"),
         am.onlyif(has_evt(Event.tree), "src/c4/yml/detail/print.hpp"),
